+++ src/components/pane.st.css
@@ -1,0 +1,5 @@
+.root {}
+
+.title {}
+
+.container {}
