+++ src/index.tsx
@@ -1,7 +1,8 @@
 import './common/reset.st.css';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App, AppModel } from './components/app';
+import { App } from './components/app';
+import { AppModel } from './lib/app-model';
 
 const container = document.body.appendChild(document.createElement('div'));
 container.id = 'playground';
