+++ src/lib/use-force-update.tsx
@@ -1,0 +1,7 @@
+import { useCallback, useState } from 'react';
+
+export function useForceUpdate(): () => void {
+    const [, dispatch] = useState(Object.create(null));
+    const memoizedDispatch = useCallback((): void => dispatch(Object.create(null)), [dispatch]);
+    return memoizedDispatch;
+}
